+++ src/components/core/pages/HomePage/HomePage.js
@@ -6,7 +6,7 @@ import Input from '../../atoms/Input/Input'
 import DragAndDrop from '../../atoms/DragAndDrop/DragAndDrop'
 import List from '../../organisms/List/List'
 import FileItem from '../../molecules/FileItem/FileItem'
-import ProfItem from '../../molecules/ProfItem/ProfItem'
+import ProfItem from '../../organisms/ProfItem/ProfItem'
 import CheckboxMessage from '../../molecules/CheckboxMessage/CheckboxMessage'
 import { concernTrigger } from '../../../../redux/actions/concernAction'
 import { connect } from 'react-redux'
